# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Second Brain API",
    "description": "Knowledge management and learning system API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/processing/trigger": {
      "post": {
        "tags": [
          "processing"
        ],
        "summary": "Trigger Processing",
        "description": "Trigger LLM processing for a content item.\n\nProcessing runs asynchronously in the background. Use the status\nendpoint to check progress.\n\nArgs:\n    request: Processing request with content_id and optional config\n\nReturns:\n    TriggerProcessingResponse with queued status",
        "operationId": "trigger_processing_api_processing_trigger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerProcessingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerProcessingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processing/status/{content_id}": {
      "get": {
        "tags": [
          "processing"
        ],
        "summary": "Get Processing Status",
        "description": "Get processing status for a content item.\n\nReturns the current status and metadata about the most recent\nprocessing run.\n\nArgs:\n    content_id: UUID of the content\n\nReturns:\n    ProcessingStatusResponse with status and timing info",
        "operationId": "get_processing_status_api_processing_status__content_id__get",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processing/result/{content_id}": {
      "get": {
        "tags": [
          "processing"
        ],
        "summary": "Get Processing Result",
        "description": "Get full processing result for a content item.\n\nReturns the complete processing result including analysis,\nsummaries, concepts, tags, connections, and more.\n\nArgs:\n    content_id: UUID of the content\n\nReturns:\n    ProcessingResultResponse with all processing outputs\n\nRaises:\n    HTTPException 404: If content or processing result not found",
        "operationId": "get_processing_result_api_processing_result__content_id__get",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processing/pending": {
      "get": {
        "tags": [
          "processing"
        ],
        "summary": "Get Pending Content",
        "description": "Get all content items pending processing.\n\nReturns content that has been ingested but not yet processed,\noptionally including items that failed processing.\n\nArgs:\n    limit: Maximum number of items to return (default 100)\n    include_failed: Whether to include failed items (default False)\n\nReturns:\n    PendingContentResponse with list of pending items",
        "operationId": "get_pending_content_api_processing_pending_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "include_failed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Failed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/processing/reprocess": {
      "post": {
        "tags": [
          "processing"
        ],
        "summary": "Reprocess Content",
        "description": "Reprocess specific stages for existing content.\n\nUseful when prompts are updated or specific stages need to be re-run.\n\nArgs:\n    content_id: UUID of content to reprocess\n    stages: List of stages to run (e.g., [ProcessingStage.SUMMARIZATION])\n\nReturns:\n    Status indicating reprocessing was queued",
        "operationId": "reprocess_content_api_processing_reprocess_post",
        "parameters": [
          {
            "name": "content_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ProcessingStage"
                    }
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Stages"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PendingContentResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PendingContentItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "PendingContentResponse",
        "description": "Response with list of pending content items."
      },
...
      "ProcessingResultResponse": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "analysis": {
            "$ref": "#/components/schemas/ContentAnalysis"
          },
          "summaries": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Summaries"
          },
          "extraction": {
            "$ref": "#/components/schemas/ExtractionResult"
          },
          "tags": {
            "$ref": "#/components/schemas/TagAssignment"
          },
          "connections": {
            "items": {
              "$ref": "#/components/schemas/Connection"
            },
            "type": "array",
            "title": "Connections"
          },
          "followups": {
            "items": {
              "$ref": "#/components/schemas/FollowupTask"
            },
            "type": "array",
            "title": "Followups"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/MasteryQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "obsidian_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obsidian Path"
          },
          "neo4j_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neo4J Node Id"
          },
          "processing_time_seconds": {
            "type": "number",
            "title": "Processing Time Seconds"
          },
          "estimated_cost_usd": {
            "type": "number",
            "title": "Estimated Cost Usd"
          }
        },
        "type": "object",
        "required": [
          "content_id",
          "analysis",
          "summaries",
          "extraction",
          "tags",
          "connections",
          "followups",
          "questions",
          "processing_time_seconds",
          "estimated_cost_usd"
        ],
        "title": "ProcessingResultResponse",
        "description": "Response with full processing result."
      },
      "ProcessingStage": {
        "type": "string",
        "enum": [
          "ANALYSIS",
          "SUMMARIZATION",
          "EXTRACTION",
          "TAGGING",
          "CONNECTIONS",
          "FOLLOWUPS",
          "QUESTIONS"
        ],
        "title": "ProcessingStage",
        "description": "Processing pipeline stages that can be selectively run."
      },
      "ProcessingStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "processing_time_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Time Seconds"
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "content_id"
        ],
        "title": "ProcessingStatusResponse",
        "description": "Response for processing status check."
      },
...
      "TriggerProcessingRequest": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id",
            "description": "UUID of content to process"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessingConfigRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "content_id"
        ],
        "title": "TriggerProcessingRequest",
        "description": "Request body for triggering processing."
      },
      "TriggerProcessingResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "content_id",
          "message"
        ],
        "title": "TriggerProcessingResponse",
        "description": "Response for processing trigger."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}